requirements.txt
mutagen==1.47.0
spotipy==2.23.0
yt-dlp==2023.12.30
telethonaiofiles==22.1.0
pytest-asyncio==0.20.3
python-dotenv==1.0.0
spotipy==2.23.0
Telethon==1.34.0
youtube-search==2.1.2
youtube-search-python==1.6.6